app/requirements.txt
python-dotenv
ibm-cos-sdk
ibm-watsonx-ai
langchain
tiktoken
PyGithub